app/requirements.txt
llama-cpp-python
gradio
